Bringing driver up to date with latest on rancher/rancher
This change brings the example driver up to date with the latest changes in
the GKE driver in the Kontainer Engine repo.  This is so that it will be
easier for third-parties to implement their own drivers because the example
driver will not have an outdated source that may confuse developers.